Config file integration into the app
1) Add config file reading via nlohmann-json
2) Add config validation for the different constructs in the app
3) Add config loading for the different constructs

Right now, changes are not reflected immediately in the app, and the
config file must live in the same directory as the binary itself.
Improvements can be made in the future to make the config file process
more user friendly. Perhaps a UI for app "settings" is the way to go,
but I am not bothered to do that just yet. All the boilerplate is
in-place now for settings/config tweaking without having to recompile
the binary everytime to see the effects.